# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ccv API",
    "description": "Claude Code Environment Manager API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/envs/{name}/skills": {
      "get": {
        "summary": "List skills",
        "tags": ["Skills"],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of skills",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "data": {
                      "$ref": "#/components/schemas/SkillListResponse"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Install skill from uploaded zip archive",
        "tags": ["Skills"],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Skill added",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "data": {
                      "$ref": "#/components/schemas/SkillInfo"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/envs/{name}/skills/{skill}": {
      "delete": {
        "summary": "Delete skill",
        "tags": ["Skills"],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skill",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Skill deleted"
          },
          "404": {
            "description": "Skill not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SkillInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "SkillListResponse": {
        "type": "object",
        "properties": {
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillInfo"
            }
          }
        }
      },
...
    }
  }
}